+++ app/src/main/java/com/quangtkd/kotlintest/fragment/FragHome.kt
@@ -53,7 +53,7 @@ class FragHome : Fragment() {
                             fiveDayWeather = it.body()
                             var celsius = ceil(fiveDayWeather?.list?.first()?.main?.temp!! - 273.15)
                             tv_frag_home_city.text = fiveDayWeather?.city?.name
-                            tv_frag_home_temp.text = celsius.toString()  +"\u00B0"
+                            tv_frag_home_temp.text = celsius.toInt().toString()  +"\u00B0"
                             WeatherArray.addAll(ArrayList(fiveDayWeather?.list))
 
                             adapter.notifyDataSetChanged()
