+++ app/src/main/java/com/quangtkd/kotlintest/fragment/FragInvestNews.kt
@@ -55,7 +55,7 @@ class FragInvestNews : Fragment() {
                             fiveDayWeather = it.body()
                             var celsius = ceil(fiveDayWeather?.list?.first()?.main?.temp!! - 273.15)
                             tv_frag_invest_news_city.text = fiveDayWeather?.city?.name
-                            tv_frag_invest_news_temp.text = celsius.toString()  +"\u00B0"
+                            tv_frag_invest_news_temp.text = celsius.toInt().toString()  +"\u00B0"
                             WeatherArray.addAll(ArrayList(fiveDayWeather?.list))
 
                             adapter.notifyDataSetChanged()
